backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
paho-mqtt
python-jose
python-multipart
bcrypt
sphinx
sphinx-rtd-theme 
myst-parser